# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Research Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Paper {
  id: string;
  title: string;
  authors: string[];
  /** @nullable */
  year?: number | null;
  /** @nullable */
  journal?: string | null;
  /** @nullable */
  abstract?: string | null;
  /** @nullable */
  doi?: string | null;
  /** @nullable */
  citationCount?: number | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  url?: string | null;
  source: string;
  hasDataset: boolean;
  /** @nullable */
  githubUrl?: string | null;
  /**
     * beginner | intermediate | advanced
     * @nullable
     */
  difficultyLevel?: string | null;
  /** @nullable */
  qualityScore?: number | null;
  isOpenAccess: boolean;
  fields: string[];
}

export interface PaperSearchResult {
  papers: Paper[];
  total: number;
  query: string;
}

export interface TrendingTopic {
  topic: string;
  paperCount: number;
  /** @nullable */
  growthRate?: number | null;
  /** @nullable */
  description?: string | null;
}

export interface TrendingResult {
  topics: TrendingTopic[];
  recentPapers: Paper[];
  topConferences: string[];
}

export interface CompareInput {
  papers: Paper[];
}

export interface ComparisonRow {
  paperId: string;
  title: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  authors?: string | null;
  /** @nullable */
  journal?: string | null;
  /** @nullable */
  dataset?: string | null;
  /** @nullable */
  algorithm?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  accuracy?: string | null;
  /** @nullable */
  citationCount?: number | null;
  /** @nullable */
  advantages?: string | null;
  /** @nullable */
  limitations?: string | null;
  /** @nullable */
  novelContribution?: string | null;
}

export interface CompareResult {
  rows: ComparisonRow[];
  aiRecommendation: string;
  /** @nullable */
  bestForBeginners?: string | null;
  /** @nullable */
  bestForImplementation?: string | null;
  /** @nullable */
  bestForFutureResearch?: string | null;
  /** @nullable */
  mostCited?: string | null;
  /** @nullable */
  newest?: string | null;
}

export interface GapInput {
  papers: Paper[];
  topic: string;
}

export interface GapResult {
  commonTechniques: string[];
  frequentDatasets: string[];
  commonLimitations: string[];
  researchGaps: string[];
  futureWork: string[];
  novelIdeas: string[];
}

export interface RoadmapInput {
  topic: string;
  papers?: Paper[];
}

export interface RoadmapStep {
  level: string;
  title: string;
  description: string;
  items: string[];
  order: number;
}

export interface RoadmapResult {
  topic: string;
  steps: RoadmapStep[];
  /** @nullable */
  estimatedTime?: string | null;
}

export interface ExplainInput {
  paper: Paper;
}

export interface ExplainResult {
  summary: string;
  objective: string;
  problem: string;
  methodology: string;
  algorithms: string[];
  /** @nullable */
  dataset?: string | null;
  /** @nullable */
  architecture?: string | null;
  metrics: string[];
  results: string;
  advantages: string[];
  limitations: string[];
  futureScope: string;
  realWorldApplications: string[];
}

export interface CitationInput {
  paper: Paper;
}

export interface CitationResult {
  apa: string;
  ieee: string;
  mla: string;
  chicago: string;
  bibtex: string;
}

export interface RelatedInput {
  paper: Paper;
  limit?: number;
}

export interface Dataset {
  id: string;
  name: string;
  description: string;
  /** @nullable */
  size?: string | null;
  format: string[];
  source: string;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  kaggleUrl?: string | null;
  /** @nullable */
  huggingfaceUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  paperCount?: number | null;
  tasks: string[];
}

export interface DatasetSearchResult {
  datasets: Dataset[];
  total: number;
}

export interface GithubRepo {
  id: string;
  name: string;
  fullName: string;
  /** @nullable */
  description?: string | null;
  stars: number;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  framework?: string | null;
  lastUpdated: string;
  url: string;
  topics: string[];
}

export interface GithubSearchResult {
  repos: GithubRepo[];
  total: number;
}

export interface ChatMessage {
  /** user | assistant */
  role: string;
  content: string;
}

export interface ChatInput {
  message: string;
  history?: ChatMessage[];
  papers?: Paper[];
}

export interface ChatResult {
  response: string;
  suggestedQuestions: string[];
}

export interface SearchHistory {
  id: number;
  query: string;
  resultCount: number;
  searchedAt: string;
}

export interface SearchHistoryResult {
  history: SearchHistory[];
}

export interface Bookmark {
  id: number;
  paperId: string;
  title: string;
  authors: string[];
  /** @nullable */
  year?: number | null;
  /** @nullable */
  url?: string | null;
  source: string;
  savedAt: string;
}

export interface BookmarkListResult {
  bookmarks: Bookmark[];
}

export interface BookmarkInput {
  paperId: string;
  title: string;
  authors: string[];
  /** @nullable */
  year?: number | null;
  /** @nullable */
  url?: string | null;
  source: string;
}

export type SearchPapersParams = {
/**
 * Research title, keywords, or problem statement
 */
query: string;
limit?: number;
/**
 * Comma-separated list of sources (semanticscholar,arxiv,openalex,crossref)
 */
sources?: string;
};

export type GetTrendingPapersParams = {
field?: string;
};

export type SearchDatasetsParams = {
query: string;
limit?: number;
};

export type SearchGithubReposParams = {
query: string;
limit?: number;
};

export type GetSearchHistoryParams = {
limit?: number;
};