# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Research Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers/compare:
    post:
      operationId: comparePapers
      tags: [ai]
      summary: AI-powered comparison of multiple papers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareInput"
      responses:
        "200":
          description: Comparison result with table and AI recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareResult"
  /papers/gaps:
    post:
      operationId: detectResearchGaps
      tags: [ai]
      summary: Detect research gaps from a list of papers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GapInput"
      responses:
        "200":
          description: Research gaps analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GapResult"
  /papers/roadmap:
    post:
      operationId: generateRoadmap
      tags: [ai]
      summary: Generate a research roadmap for a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapInput"
      responses:
        "200":
          description: Research roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapResult"
  /papers/explain:
    post:
      operationId: explainPaper
      tags: [ai]
      summary: Generate a student-friendly AI explanation for a paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainInput"
      responses:
        "200":
          description: Paper explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainResult"
...
  /datasets/search:
    get:
      operationId: searchDatasets
      tags: [datasets]
      summary: Search for datasets related to a research topic
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetSearchResult"
...
components:
  schemas:
...
    Paper:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: ["integer", "null"]
        journal:
          type: ["string", "null"]
        abstract:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        citationCount:
          type: ["integer", "null"]
        pdfUrl:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        source:
          type: string
        hasDataset:
          type: boolean
        githubUrl:
          type: ["string", "null"]
        difficultyLevel:
          type: ["string", "null"]
          description: "beginner | intermediate | advanced"
        qualityScore:
          type: ["number", "null"]
        isOpenAccess:
          type: boolean
        fields:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - authors
        - source
        - isOpenAccess
        - hasDataset
        - fields
...
    CompareInput:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
      required:
        - papers
    ComparisonRow:
      type: object
      properties:
        paperId:
          type: string
        title:
          type: string
        year:
          type: ["integer", "null"]
        authors:
          type: ["string", "null"]
        journal:
          type: ["string", "null"]
        dataset:
          type: ["string", "null"]
        algorithm:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        accuracy:
          type: ["string", "null"]
        citationCount:
          type: ["integer", "null"]
        advantages:
          type: ["string", "null"]
        limitations:
          type: ["string", "null"]
        novelContribution:
          type: ["string", "null"]
      required:
        - paperId
        - title
    CompareResult:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ComparisonRow"
        aiRecommendation:
          type: string
        bestForBeginners:
          type: ["string", "null"]
        bestForImplementation:
          type: ["string", "null"]
        bestForFutureResearch:
          type: ["string", "null"]
        mostCited:
          type: ["string", "null"]
        newest:
          type: ["string", "null"]
      required:
        - rows
        - aiRecommendation
    GapInput:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
        topic:
          type: string
      required:
        - papers
        - topic
    GapResult:
      type: object
      properties:
        commonTechniques:
          type: array
          items:
            type: string
        frequentDatasets:
          type: array
          items:
            type: string
        commonLimitations:
          type: array
          items:
            type: string
        researchGaps:
          type: array
          items:
            type: string
        futureWork:
          type: array
          items:
            type: string
        novelIdeas:
          type: array
          items:
            type: string
      required:
        - commonTechniques
        - frequentDatasets
        - commonLimitations
        - researchGaps
        - futureWork
        - novelIdeas
    RoadmapInput:
      type: object
      properties:
        topic:
          type: string
        papers:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
      required:
        - topic
    RoadmapStep:
      type: object
      properties:
        level:
          type: string
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            type: string
        order:
          type: integer
      required:
        - level
        - title
        - description
        - items
        - order
    RoadmapResult:
      type: object
      properties:
        topic:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapStep"
        estimatedTime:
          type: ["string", "null"]
      required:
        - topic
        - steps
    ExplainInput:
      type: object
      properties:
        paper:
          $ref: "#/components/schemas/Paper"
      required:
        - paper
    ExplainResult:
      type: object
      properties:
        summary:
          type: string
        objective:
          type: string
        problem:
          type: string
        methodology:
          type: string
        algorithms:
          type: array
          items:
            type: string
        dataset:
          type: ["string", "null"]
        architecture:
          type: ["string", "null"]
        metrics:
          type: array
          items:
            type: string
        results:
          type: string
        advantages:
          type: array
          items:
            type: string
        limitations:
          type: array
          items:
            type: string
        futureScope:
          type: string
        realWorldApplications:
          type: array
          items:
            type: string
      required:
        - summary
        - objective
        - problem
        - methodology
        - algorithms
        - metrics
        - results
        - advantages
        - limitations
        - futureScope
        - realWorldApplications
...
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        size:
          type: ["string", "null"]
        format:
          type: array
          items:
            type: string
        source:
          type: string
        downloadUrl:
          type: ["string", "null"]
        kaggleUrl:
          type: ["string", "null"]
        huggingfaceUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        paperCount:
          type: ["integer", "null"]
        tasks:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - source
        - format
        - tasks
    DatasetSearchResult:
      type: object
      properties:
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/Dataset"
        total:
          type: integer
      required:
        - datasets
        - total
...