# lib/api-zod/src/generated/types/compareResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Research Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ComparisonRow } from './comparisonRow';

export interface CompareResult {
  rows: ComparisonRow[];
  aiRecommendation: string;
  /** @nullable */
  bestForBeginners?: string | null;
  /** @nullable */
  bestForImplementation?: string | null;
  /** @nullable */
  bestForFutureResearch?: string | null;
  /** @nullable */
  mostCited?: string | null;
  /** @nullable */
  newest?: string | null;
}