# lib/api-zod/src/generated/types/roadmapResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Research Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapStep } from './roadmapStep';

export interface RoadmapResult {
  topic: string;
  steps: RoadmapStep[];
  /** @nullable */
  estimatedTime?: string | null;
}